requirements.txt
# Core Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Database & Memory
chromadb>=0.4.0
# sqlite3 is built-in with Python

# Ollama Integration
ollama>=0.4.0
requests>=2.31.0

# Agent Framework (Microsoft)
# Note: Install when available or use alternative like autogen
# microsoft-agents

# Model Context Protocol (MCP)
# Note: MCP client to be added when official package available
# For now, we'll use direct API calls

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0
fastapi>=0.111.0
uvicorn>=0.27.0
openai>=1.66.0
tiktoken>=0.9.0
tenacity>=9.0.0
loguru>=0.7.3
structlog>=24.0.0
boto3>=1.34.0
browser-use>=0.1.40
playwright>=1.51.0

# Development
pytest>=8.0.0
black>=24.0.0
